backend/requirements.txt
flask
flask-cors
google-generativeai
requests
beautifulsoup4
python-dotenv
gunicorn
